+++ electron.vite.config.ts
@@ -1,0 +1,44 @@
+import { resolve } from 'node:path';
+import { defineConfig, externalizeDepsPlugin } from 'electron-vite';
+import react from '@vitejs/plugin-react';
+
+const sharedAlias = {
+  '@shared': resolve(__dirname, 'src/shared'),
+};
+
+export default defineConfig({
+  resolve: {
+    alias: sharedAlias,
+  },
+  main: {
+    plugins: [externalizeDepsPlugin({ exclude: [] })],
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve(__dirname, 'src/main/index.ts'),
+        },
+      },
+    },
+  },
+  preload: {
+    plugins: [externalizeDepsPlugin()],
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve(__dirname, 'src/preload/index.ts'),
+        },
+      },
+    },
+  },
+  renderer: {
+    root: resolve(__dirname, 'src/renderer'),
+    build: {
+      rollupOptions: {
+        input: {
+          index: resolve(__dirname, 'src/renderer/index.html'),
+        },
+      },
+    },
+    plugins: [react()],
+  },
+});